pyproject.toml
[project]
name = "semantiva"
dynamic = ["version"]
description = "Semantiva: An HPC-ready, domain-driven, type-oriented framework that delivers semantic transparency to advanced scientific computing."
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "black >=24.10.0",
    "mypy >=1.14.1",
    "numpy >=2.2.1",
    "pylint >=3.3.3",
    "pytest >=8.3.4",
    "pyyaml",
    "types-PyYAML",
    "coverage",
    "rdflib",
]
distribution = true

[tool.black]
# Configuration for the black code formatter


[tool.pylint]
# Configuration for pylint (static analysis)
max-line-length = 120  
fail-under = 7.5

[tool.mypy]
exclude = [
    "^tests/test_string_specialization\\.py$",
    "docs/", 
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["semantiva"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
cicheck = "bash ./ci_pipeline.sh"  # Define the shell script as a PDM script

[tool.pdm.version]
source = "scm"
fallback_version = "0.1.0-dev0+gUNKNOWN"

[tool.setuptools]
include = ["NOTICE", "LICENSE", "README.md"]